SVM: Remove accounts-db deps in accounts_loader tests
