+++ accounts-db/src/lib.rs
@@ -40,7 +40,6 @@ pub mod sorted_storages;
 pub mod stake_rewards;
 pub mod storable_accounts;
 pub mod tiered_storage;
-pub mod transaction_results;
 pub mod utils;
 mod verify_accounts_hash_in_background;
 pub mod waitable_condvar;
