+++ banks-server/Cargo.toml
@@ -13,12 +13,12 @@ edition = { workspace = true }
 bincode = { workspace = true }
 crossbeam-channel = { workspace = true }
 futures = { workspace = true }
-solana-accounts-db = { workspace = true }
 solana-banks-interface = { workspace = true }
 solana-client = { workspace = true }
 solana-runtime = { workspace = true }
 solana-sdk = { workspace = true }
 solana-send-transaction-service = { workspace = true }
+solana-svm = { workspace = true }
 tarpc = { workspace = true, features = ["full"] }
 tokio = { workspace = true, features = ["full"] }
 tokio-serde = { workspace = true, features = ["bincode"] }
