+++ runtime/src/bank_utils.rs
@@ -7,8 +7,8 @@ use {
     solana_sdk::{pubkey::Pubkey, signature::Signer},
 };
 use {
-    solana_accounts_db::transaction_results::TransactionResults,
     solana_sdk::transaction::SanitizedTransaction,
+    solana_svm::transaction_results::TransactionResults,
     solana_vote::{vote_parser, vote_sender_types::ReplayVoteSender},
 };
