+++ svm/src/lib.rs
@@ -8,6 +8,7 @@ pub mod runtime_config;
 pub mod transaction_account_state_info;
 pub mod transaction_error_metrics;
 pub mod transaction_processor;
+pub mod transaction_results;
 
 #[macro_use]
 extern crate solana_metrics;
