+++ svm/src/transaction_processor.rs
@@ -1,19 +1,18 @@
 use {
     crate::{
-        account_loader::{load_accounts, TransactionCheckResult},
+        account_loader::{
+            load_accounts, LoadedTransaction, TransactionCheckResult, TransactionLoadResult,
+        },
         account_overrides::AccountOverrides,
         runtime_config::RuntimeConfig,
         transaction_account_state_info::TransactionAccountStateInfo,
         transaction_error_metrics::TransactionErrorMetrics,
-    },
-    log::debug,
-    percentage::Percentage,
-    solana_accounts_db::{
-        accounts::{LoadedTransaction, TransactionLoadResult},
         transaction_results::{
             DurableNonceFee, TransactionExecutionDetails, TransactionExecutionResult,
         },
     },
+    log::debug,
+    percentage::Percentage,
     solana_measure::measure::Measure,
     solana_program_runtime::{
         compute_budget::ComputeBudget,
